Overwrite the IDs for parameters.
In order to run parametrized tests in multiple threads, it is necessary to make sure that the IDs are unique for all threads. Since `fauxfactory` will generate random values, each thread being spawn will receive a different set of IDs, which does not work.

This PR will overwrite the IDs by using the name of the mark being used and appending a numeric index to it.